add position to board and relocate api crud to separate modules
